% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyNifty.r
\name{applyNifty}
\alias{applyNifty}
\title{Apply NiftyReg transform found through registration on entire "MSImageSet"}
\usage{
applyNifty(cardinaldata, niftyregtform, padding = 0, interpolation = 0)
}
\arguments{
\item{cardinaldata}{an object of class "MSImageSet"}

\item{niftyregtform}{the transformation developed during the RNiftyReg registration step}

\item{padding}{zero padding used around the borders of the image to avoid out of frame pixels}

\item{interpolation, }{passed to RNiftyReg. Defaults to nearest neighbor}
}
\value{
cardinaldata_registered a registered "MSImageSet"
}
\description{
Applies the NiftyReg transform on the entire MSIimageSet iteratively
}
