% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundBox.r
\name{boundBox}
\alias{boundBox}
\title{Bounding box of zeros around Cardinaldata}
\usage{
boundBox(cardinaldata, xmax = 300, ymax = 250, center = F)
}
\arguments{
\item{cardinaldata}{an object of class "MSImageSet"}

\item{xmax}{maximum of bounding box in x}

\item{ymax}{maximum of bounding box in y}

\item{boolean, }{whether to center the imaging data in the bounding box}
}
\value{
msset an object of class "MSImageSet"
}
\description{
Adds zeros around "MSImageSet" object to so that datasets can be combined after registration
}
