% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineReggedIMS.R
\name{combineReggedIMS}
\alias{combineReggedIMS}
\title{Combine registered IMS datasets into a single format}
\usage{
combineReggedIMS(cardinaldata1, cardinaldata2, ds1_name = "dataset1",
  ds2_name = "dataset2", combined_name = "combined_ds")
}
\arguments{
\item{cardinaldata1}{an object of class "MSImageSet", first dataset}

\item{cardinaldata2}{an object of class "MSImageSet", second dataset}

\item{ds1_name}{name added to the data under @featureData$ds_origin for first dataset}

\item{combined_name}{passed as 'sample' in resultant "MSImageSet"}

\item{ds1_name}{name added to the data under @featureData$ds_origin for second dataset}
}
\value{
a combined  "MSImageSet"
}
\description{
Combines datasets with an mz_offset (soon:or with a featureData label)
}
