% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationQuery.R
\name{correlationQuery}
\alias{correlationQuery}
\title{Query correlation matrix for top markers accross combined data sets}
\usage{
correlationQuery(combined_cardinaldata, correlation_matrix,
  query_dataset = "mydata", query_against = c("otherdata", "all"),
  query_mz = 885.55, plot_ions = T, top_n = 9, ...)
}
\arguments{
\item{combined_cardinaldata}{an object of class "MSImageSet", combined dataset}

\item{correlation_matrix}{object generated from getCorrelationMatrix}

\item{query_dataset}{name of the dataset in 'ds_origin' that we wish to query for}

\item{query_against}{name of the dataset in 'ds_origin' that we will query against, can be multiple, passing 'all' will query all datasets}

\item{query_mz}{mz in the query_dataset one wishes to find correlations to in query_against dataset(s)}

\item{plot_ions}{boolean, whether to make a plot of top ions}

\item{top_n}{number of ions to plot}

\item{...}{arguments passed to image, useful for passing layout=c() of image with multiple queries}
}
\value{
data.frame of mz and r, sorted in descending order by r for the query_mz, first value is always mz_query
}
