% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationMat.R
\name{getCorrelationMat}
\alias{getCorrelationMat}
\title{Get correlation matrix from Cardinal dataset}
\usage{
getCorrelationMat(combined_cardinaldata)
}
\arguments{
\item{combined_cardinaldata}{an object of class "MSImageSet", first dataset}
}
\value{
a correlation matrix with row and columns names as mzs
}
\description{
Makes a correlation matrix using base R's 'cor' function, names columns and rows mz
}
