% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceCardinalCoord.r
\name{reduceCardinalCoord}
\alias{reduceCardinalCoord}
\title{Reduce Cardinal dataset's coordinates to minimum}
\usage{
reduceCardinalCoord(cardinaldata)
}
\arguments{
\item{cardinaldata}{an object of class "MSImageSet"}
}
\value{
cardinaldata
}
\description{
reduceCardinalCoord is used to set the minimum y and x coordinate values in a Cardinal dataset to zero.
}
