% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultToArray.r
\name{resultToArray}
\alias{resultToArray}
\title{Cardinal resultSet to R array}
\usage{
resultToArray(cardinaldata, resultdata, idx)
}
\arguments{
\item{cardinaldata}{an object of class "MSImageSet"}

\item{resultdata}{an object of class "ResultSet"}
}
\value{
image_matrix
}
\description{
converts a Cardinal ResultSet data (stored as 1D vector) to 2D array, for now, only works for SpatialShrunkenCentroids and PCA
}
