% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedReggedIMS.r
\name{combineReggedIMS}
\alias{combineReggedIMS}
\title{Combine registered IMS datasets into a single format}
\usage{
combineReggedIMS(cardinaldata1, cardinaldata2, mz_offset = 13000,
  sample_name = "combined_ds")
}
\arguments{
\item{cardinaldata1}{an object of class "MSImageSet", first dataset}

\item{cardinaldata2}{an object of class "MSImageSet", second dataset}

\item{mz_offset}{offset of the m/z axis between datasets}

\item{sample_name}{passed as 'sample' in resultant "MSImageSet"}
}
\value{
a combined offset "MSImageSet"
}
\description{
Combines datasets with an mz_offset (soon:or with a featureData label)
}
